Refactor component structure: move CardProfile to a new directory and restore BackgroundGrid component
